pyproject.toml
[project]
name = "online-bs"
version = "0.1.0"
description = "Exploring batch selection methods for accelerating the training of machine learning models."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "clip",
    "cvxpy>=1.7.2",
    "ema-pytorch>=0.7.7",
    "ftfy>=6.3.1",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "matplotlib>=3.10.6",
    "numpy>=2.2.6",
    "pillow>=11.3.0",
    "pre-commit>=4.2.0",
    "pyyaml>=6.0.2",
    "regex>=2025.7.33",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "scipy>=1.15.3",
    "fiftyone>=0.25.0",
    "opencv-python>=4.8.0",
    "timm>=1.0.24",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "transformers>=5.0.0",
    "umap-learn>=0.5.5",
    "wandb>=0.21.0",
    "wcwidth>=0.2.13",
    "pytest>=9.0.3",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }

# CLIP is based on old architecture, which results in a setup bug
# this fixes that bug, because it uses a version of setuptools
# that fixes this issue
[tool.uv]
build-constraint-dependencies = ["setuptools<81"]
